feat: 좋아요한 트랙 목록 조회

+++ frontend/app/src/main/java/com/whistlehub/common/view/navigation/AppContentNavGraph.kt
@@ -68,8 +68,8 @@ fun AppContentNavGraph(
                 navController
             )
         }
-        composable(route = Screen.Profile.route + "/{memberId}"){ backStackEntry ->
-        val memberId = backStackEntry.arguments?.getString("memberId")
+        composable(route = Screen.Profile.route + "/{memberId}") { backStackEntry ->
+            val memberId = backStackEntry.arguments?.getString("memberId")
             ProfileScreen(
                 memberId?.toInt() ?: -1,
                 navController = navController,
@@ -105,7 +105,7 @@ fun AppContentNavGraph(
             if (playlistId != null) {
                 PlaylistTrackListScreen(
                     paddingValues,
-                    playlistId.toInt(),
+                    playlistId,
                     navController,
                     trackPlayViewModel
                 )
